pyproject.toml
[tool.poetry]
name = "real-estate-api"
version = "0.1.0"
description = ""
authors = ["Theodore Flokos <t.flokos@nvisionist.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.4"
djangorestframework = "^3.16.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.1"
djangorestframework-simplejwt = "^5.5.1"
gunicorn = "^23.0.0"
django-filter = "^25.1"
drf-yasg = "^1.21.10"
whitenoise = "^6.6.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"
coverage = "^7.10.1"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
